Move test with digital product and checkoutComplete to test_checkout_complete. Fix incorrect mock patchess
